secp256k1: Move curve benchmarks to own file.
This splits the benchmarks related to curve operations into their own
file to be more consistent with other benchmarks in the package.
